\docType{data}
\name{pdp133}
\alias{pdp133}
\title{Garnet Lu-Hf Sample PdP13-3 from Mulcahy et al., 2011.}
\format{A data frame with 7 rows and 5 variables}
\description{
  A dataset containing the geochemical data for reproducing
  the isochron for sample PdP13-3 in Figure 15A of Mulcahy
  et al., 2011:
}
\details{
  \itemize{ \item labels Mineral phase analyzed \item x
  \eqn{^{176}}Lu/\eqn{^{177}}Hf \item sx 1-sigma
  uncertainty on \eqn{^{176}}Lu/\eqn{^{177}}Hf \item y
  \eqn{^{176}}Hf/\eqn{^{177}}Hf \item sy 1-sigma
  uncertainty on \eqn{^{176}}Lu/\eqn{^{177}}Hf ... }
}
\references{
  Mulcahy, S. R., Roeske, S. M., McClelland, W. C.,
  Jourdan, F., Iriondo, A., Renne, P. R., Vervoort, J. D.,
  and Vujovich, G. I., 2011, Structural evolution of a
  composite middle to lower crustal section: The Sierra de
  Pie de Palo, northwest Argentina: Tectonics, v. 30, p.
  TC1005.
  \url{http://www.agu.org/pubs/crossref/2011/2009TC002656.shtml}
}
\keyword{datasets}

