\docType{data}
\name{russel}
\alias{russel}
\title{Pb-Pb isochron data of Russel, 1995.}
\format{A data frame with 26 rows and 5 variables}
\description{
  A dataset containing the geochemical data for reproducing
  the isochron in Figure 8 of Powell, 2002 using data from
  Russel, 1995:
}
\details{
  \itemize{ \item labels Point labels \item x
  \eqn{^{206}}Pb/\eqn{^{204}}Pb \item sx 1-sigma
  uncertainty on \eqn{^{206}}Pb/\eqn{^{204}}Pb \item y
  \eqn{^{207}}Pb/\eqn{^{204}}Pb \item sy 1-sigma
  uncertainty on \eqn{^{207}}Pb/\eqn{^{204}}Pb }
}
\references{
  Russell, J., 1995. Direct dating of Silurian macrofossils
  from Gotland, Sweden. In: Dunay, R.E., Hailwood, E.A.
  (Eds.), Non- Biostratigraphical Methods of Dating and
  Correlation. Geological Society Special Publication, vol.
  89, pp. 175-200.
  \url{http://sp.lyellcollection.org/content/
  89/1/175.abstract}
}
\keyword{datasets}

